myrecord.pyをapp.pyに統合。create_tables_MySQL.pyをbiclove_flowledge用に変更
